pyproject.toml
[project]
name = "nsf-osdf-map"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "elasticsearch>=9.2.1",
    "python-dotenv>=1.2.1",
]
